fix: Use the matching "stop" icon for an item card
When an item card does **not** have the timer running, the icon is the
triangular play button, inside of a circle. When the timer is started,
the button's icon is the square "stop" button. However, this is not
contained within a circle. This looks strange by comparison, and looks
like a bug (as it's just a white box).

Fix this by using the `fa-stop-circle` icon, which matches the existing
`fa-play-circle` button.