Update README.md
I made two updates:  (1) I inserted an image to show were to click on "Team Assessment" link at the top of Section 7 and (2) I inserted a new Best Practice at the beginning of the "Best Practice" section - titled "Plan for a Team Assessment Retrospective" - which includes 2 tables. First table is what to do before the event and second table is to show time boxes for hosting the event.